* Reworked link handling and generation
Bifrost is now able to generate stable uuids for Light and GroupedLight. This
prevents the state database from growing with boundless copies of each detected
object, and makes the hue app able to understand (as of yet, unsupported)
incoming state updates.

The uuid generation still needs more work, but this is a viable starting point.

This marks the first version that is able to show z2m groups as rooms, each with
their corresponding lights, and send updates back the hue app!