+ Implemented scene recall learning
One fundamental problem with interfacing z2m and hue, is that some information
needed for hue is simply not present (or available) in z2m.

For example, a "Scene" in hue::v2 should contain settings for all lights in the
scene (brightness, color, etc), while z2m only knows (index, name) for a scene.

With this new code, bifrost will try to learn missing information from scenes as
they are being recalled (activated).

When a scene is activated, the first update to each light within 3 seconds will
be recorded as the correct state for that scene.

This isn't a perfect method, as any out-of-order updates from z2m in the same
time window might be recorded as the intended scene setting, but it sure beats
having no information at all.