* Finally figured out how to discern lights in "ct" vs "xy" mode!
The color_temperature.mirek field can be null! If it is, the light is not in ct
mode (and thus, will be in xy mode, if supported, I believe).

This finally makes sense, and we can remove the hackish Light.color_mode field,
that was used to keep track of exactly this state.